feat: add main operation system on pressing equal
Created functions to store the operator and compute the right operation
on pressing equal button based on operator sign (setOperator() &
operate() functions). For that, added IDs to HTML elements.

Refactored display functions which weren't taking into account the
difference between numeric variables and strings relative to variables
displayed on screen, and added updateDisplay() function to move current
display to previous display upon choosing an operator.

Added parse methods to operate() to convert the a and b strings into
numbers, as well as the result, before converting the result back to a
number.

Added function to handle decimal numbers.